pipelines/scripts/cancerf/requirements.txt
absl-py==2.1.0
anndata==0.10.9
array-api-compat==1.9
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
chex==0.1.87
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.9.0
get-annotations==0.1.2
h5py==3.12.1
idna==3.7
igraph==0.11.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
legacy-api-wrap==1.4
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.5.0
natsort==8.4.0
numba==0.60.0
numpy==1.24.3
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.4.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pip==24.2
platformdirs==4.3.6
plottable==0.1.5
prompt-toolkit==3.0.48
psutil==5.9.0
pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
rich==13.9.3
scanpy==1.10.3
scib-metrics==0.5.1
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
session-info==1.0.0
setuptools==75.1.0
six==1.16.0
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
texttable==1.7.0
threadpoolctl==3.5.0
toolz==1.0.0
torchdata==0.7.1
torchtext==0.18.0
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
zipp==3.20.2